+++ Effect.h
@@ -12,16 +12,19 @@
 
 class Effect {
 public:
-	Effect(int level, const int* index, int num);
+	Effect( int level, const uint8_t* index, int num );
 	virtual ~Effect();
 
 	virtual bool updateLeds( unsigned long now, CRGB* leds ) = 0;
-	virtual void start() { startTime = millis(); active = true; }
+	virtual void start() {
+		startTime = millis();
+		active = true;
+	}
 
 	int level;
 	uint32_t startTime;
 	bool active = false;
-	const int* index;
+	const uint8_t* index;
 	int num;
 };
